move insert function to header, so we can reuse it
