+++ C/LinkedList/Singly_Linked_List/Deleting.c
@@ -1,10 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
-
-typedef struct node{
-	int data;
-	struct node *nextAddress;
-}Node;
+#include "linked_list.h"
 
 Node *start = NULL;
 
@@ -19,26 +15,6 @@ int sizeofList()
 		ptr = ptr->nextAddress;
 	}
 	return count;
-}
-void add(int no)
-{
-	Node *newNode, *ptr;
-	newNode = (Node *)malloc(sizeof(Node));
-	newNode->data = no;
-	newNode->nextAddress = NULL;
-	if(start == NULL)
-	{
-		start = newNode;
-	}
-	else
-	{
-		ptr = start;
-		while(ptr->nextAddress!=NULL)
-		{
-			ptr = ptr->nextAddress;
-		}
-		ptr->nextAddress = newNode;
-	}
 }
 void deleteBeg()
 {
@@ -92,7 +68,7 @@ int main()
 			printf("Enter Number : ");
 			scanf("%d",&no);
 			printf("\n");
-			add(no);
+			insert(&start, no);
 		}
 		else if(choice==2)
 		{
@@ -113,10 +89,10 @@ int main()
 				else
 				{
 					deletePos(pos);
-					break;				
+					break;
 				}
-			}	
-			printf("\n");	
+			}
+			printf("\n");
 		}
 		else if(choice==3)
 		{
